Move useHiddenLocation
