+++ generatedTypes/src/incubator/TransitionView/useAnimatedTransition.d.ts
@@ -1,4 +1,4 @@
-import { Direction, HiddenLocation } from './useHiddenLocation';
+import { Direction, HiddenLocation } from '../hooks/useHiddenLocation';
 import { AnimationNotifierEndProps } from './useAnimationEndNotifier';
 export interface AnimatedTransitionProps extends AnimationNotifierEndProps {
     /**
